Shift TreasuryAssetsGraph to using a custom hook for fetching TokenRecords
