+++ src/hooks/useInfiniteTokenRecords.ts
@@ -1,0 +1,82 @@
+import { useEffect, useMemo, useRef, useState } from "react";
+import {
+  TokenRecord,
+  TokenRecord_Filter,
+  TokenRecordsQuery,
+  TokenRecordsQueryVariables,
+  useInfiniteTokenRecordsQuery,
+} from "src/generated/graphql";
+import { adjustDateByDays, getISO8601String } from "src/helpers/DateHelper";
+import { DEFAULT_RECORD_COUNT } from "src/views/TreasuryDashboard/components/Graph/Constants";
+import { getNextPageStartDate } from "src/views/TreasuryDashboard/components/Graph/helpers/SubgraphHelper";
+import {
+  getNextPageParamFactory,
+  getTokenRecordDateMap,
+} from "src/views/TreasuryDashboard/components/Graph/helpers/TokenRecordsQueryHelper";
+
+export const useInfiniteTokenRecordsQueries = (
+  chartName: string,
+  subgraphUrl: string,
+  baseFilter: TokenRecord_Filter,
+  earliestDate: string | null,
+) => {
+  const initialFinishDate = getISO8601String(adjustDateByDays(new Date(), 1)); // Tomorrow
+  const initialStartDate = !earliestDate ? null : getNextPageStartDate(initialFinishDate, earliestDate, -180); // TODO restore offset
+
+  const paginator = useRef<(lastPage: TokenRecordsQuery) => TokenRecordsQueryVariables | undefined>();
+  useEffect(() => {
+    // We can't create the paginator until we have an earliestDate
+    if (!earliestDate || !baseFilter) {
+      return;
+    }
+
+    console.info(`${chartName}: earliestDate changed to ${earliestDate}. Re-fetching.`);
+
+    // Force fetching of data with the new paginator
+    // Calling refetch() after setting the new paginator causes the query to never finish
+    refetch();
+
+    // Create a new paginator with the new earliestDate
+    paginator.current = getNextPageParamFactory(chartName, earliestDate, DEFAULT_RECORD_COUNT, baseFilter);
+  }, [baseFilter, earliestDate]);
+
+  const { data, hasNextPage, fetchNextPage, refetch } = useInfiniteTokenRecordsQuery(
+    { endpoint: subgraphUrl },
+    "filter",
+    {
+      filter: {
+        ...baseFilter,
+        date_gte: initialStartDate,
+        date_lt: initialFinishDate,
+      },
+      recordCount: DEFAULT_RECORD_COUNT,
+    },
+    {
+      enabled: earliestDate !== null && baseFilter != null,
+      getNextPageParam: paginator.current,
+    },
+  );
+
+  useEffect(() => {
+    if (hasNextPage) {
+      console.debug(chartName + ": fetching next page");
+      fetchNextPage();
+      return;
+    }
+  }, [data, hasNextPage, fetchNextPage]);
+
+  const [byDateTokenRecords, setByDateTokenRecords] = useState(new Map<string, TokenRecord[]>());
+
+  useMemo(() => {
+    if (hasNextPage || !data) {
+      console.debug(`${chartName}: Removing cached data, as query is in progress.`);
+      return;
+    }
+
+    const tokenRecords = data.pages.map(query => query.tokenRecords).flat();
+    const dateTokenRecords = getTokenRecordDateMap(tokenRecords);
+    setByDateTokenRecords(dateTokenRecords);
+  }, [hasNextPage, data]);
+
+  return byDateTokenRecords;
+};
